Update add docs executor to update existing docs
