+++ language-server/modules/langserver-core/src/test/resources/codeaction/add-documentation/config/documentAlreadyDocumentedConfig1.json
@@ -1,0 +1,36 @@
+{
+  "line": 5,
+  "character": 19,
+  "expected": [
+    {
+      "title": "Document this",
+      "command": {
+        "title": "Document this",
+        "command": "ADD_DOC",
+        "arguments": [
+          {
+            "key": "node.range",
+            "value": {
+              "start": {
+                "line": 1,
+                "character": 0
+              },
+              "end": {
+                "line": 7,
+                "character": 1
+              }
+            }
+          }
+        ]
+      }
+    },
+    {
+      "title": "Document all",
+      "command": {
+        "title": "Document all",
+        "command": "ADD_ALL_DOC",
+        "arguments": []
+      }
+    }
+  ]
+}
