+++ language-server/modules/langserver-core/src/test/resources/codeaction/add-documentation/source/alreadyDocumentedSource.bal
@@ -1,0 +1,8 @@
+
+# Given the last used ID, this will return the next ID to be used
+#
+# + lastId - Last used ID
+# + return - Next ID to be used
+public function generateId(int lastId) returns int {
+    return lastId + 1;
+}
