+++ language-server/modules/langserver-core/src/test/resources/command/add-documentation/config/document_already_documented_config1.json
@@ -2,12 +2,12 @@
   "arguments": {
     "node.range": {
       "start": {
-        "line": 3,
-        "character": 4
+        "line": 1,
+        "character": 0
       },
       "end": {
-        "line": 10,
-        "character": 5
+        "line": 8,
+        "character": 1
       }
     }
   },
@@ -20,15 +20,15 @@
               {
                 "range": {
                   "start": {
-                    "line": 3,
-                    "character": 4
+                    "line": 1,
+                    "character": 0
                   },
                   "end": {
-                    "line": 6,
+                    "line": 5,
                     "character": 31
                   }
                 },
-                "newText": "# This is a test method within a class\n    #\n    # + b - Parameter Description\n    # # Deprecated\n    # This method is deprecated  \n    "
+                "newText": "# Given the last used ID, this will return the next ID to be used\n# for required purposes\n#\n# + lastId - Last used ID\n# + return - Next ID to be used  \n"
               }
             ]
           }
