+++ language-server/modules/langserver-core/src/test/resources/command/add-documentation/config/document_already_documented_config2.json
@@ -2,12 +2,12 @@
   "arguments": {
     "node.range": {
       "start": {
-        "line": 3,
-        "character": 4
+        "line": 10,
+        "character": 0
       },
       "end": {
-        "line": 10,
-        "character": 5
+        "line": 20,
+        "character": 1
       }
     }
   },
@@ -20,15 +20,15 @@
               {
                 "range": {
                   "start": {
-                    "line": 3,
-                    "character": 4
+                    "line": 10,
+                    "character": 0
                   },
                   "end": {
-                    "line": 6,
-                    "character": 31
+                    "line": 11,
+                    "character": 25
                   }
                 },
-                "newText": "# This is a test method within a class\n    #\n    # + b - Parameter Description\n    # # Deprecated\n    # This method is deprecated  \n    "
+                "newText": "# A class to perform encryption and decryption.\n# Another line of comment  \n"
               }
             ]
           }
