+++ language-server/modules/langserver-core/src/test/resources/command/add-documentation/config/document_already_documented_config3.json
@@ -2,11 +2,11 @@
   "arguments": {
     "node.range": {
       "start": {
-        "line": 3,
+        "line": 14,
         "character": 4
       },
       "end": {
-        "line": 10,
+        "line": 19,
         "character": 5
       }
     }
@@ -20,15 +20,15 @@
               {
                 "range": {
                   "start": {
-                    "line": 3,
+                    "line": 14,
                     "character": 4
                   },
                   "end": {
-                    "line": 6,
-                    "character": 31
+                    "line": 16,
+                    "character": 46
                   }
                 },
-                "newText": "# This is a test method within a class\n    #\n    # + b - Parameter Description\n    # # Deprecated\n    # This method is deprecated  \n    "
+                "newText": "# Encrypts the provided plain text\n    #\n    # + plainText - plain text to be encrypted  \n    # + padding - Parameter Description\n    # + return - Return Value Description  \n    "
               }
             ]
           }
