+++ language-server/modules/langserver-core/src/test/resources/command/update-documentation/config/updateDocumentationConfig1.json
@@ -28,7 +28,7 @@
                     "character": 18
                   }
                 },
-                "newText": "# This function has a doc with multiple lines. Once updated, it should\n# preserve these lines\n#\n# + name - My Name  \n# + val - Parameter Description"
+                "newText": "# This function has a doc with multiple lines. Once updated, it should\n# preserve these lines\n#\n# + name - My Name  \n# + val - Parameter Description  \n"
               }
             ]
           }
